tendermint-proxy: ❕ move conversions into submodule
